Allow backends to be "mandatory" and have a default value per user on a per notification basis.
It's the BackendConfiguration model that supplies additional default and
mandatory settings.

I also addressed the get_backends() inefficiencies. It should perform fewer
queries overall to make up for the fact that I added one to pull the list of
relevant BackendConfiguration objects.

There was a bug in get_backends where if a user was not a member of any
groups, the code treated that user as if it was a member of all groups. This
is fixed.

Fields that are mandatory get disabled and readonly set on their formset
fields.

Added a manage.py and completed the settings.py so that one can run the
tests out of the box with ./manage.py test (assuming that django and south
    are in one's virtual env)

I added a description field to the Notification object. It gives us a way
of explaining to the user when this notification is sent.